Integrate and add tests
