+++ .gitignore
@@ -11,3 +11,5 @@ Manifest.toml
 docs/src/tutorial/*.json
 docs/src/*/*.ipynb
 docs/src/tutorial/spaghetti_plot.html
+test/*.lattice.json*
+test/*.problem.json*
