--- src/MSPFormat/.gitignore
@@ -1,2 +1,0 @@
-*.json
-*.json.gz
